requirements.txt
asgiref==3.6.0
cffi==1.15.1
cryptography==39.0.1
dj-database-url==1.2.0
Django==4.1.7
django-cors-headers==3.13.0
django-rest-knox==4.2.0
djangorestframework==3.14.0
gunicorn==20.1.0
psycopg2-binary==2.9.5
pycparser==2.21
pytz==2022.7.1
sqlparse==0.4.3
whitenoise==6.3.0
